requirements.txt
cached-property==1.5.2
cachetools==4.2.4
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
chevron==0.14.0
frozendict==2.3.9
genanki==0.13.1
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.48.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pysrt==1.1.2
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
urllib3==2.1.0
